requirements.txt
python-dotenv 
langchain 
langchain-openai 
langchain-cohere 
pymupdf 
faiss-cpu 
cohere